backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
imageio
google-genai
python-multipart
reportlab
python-dotenv
